chatBot/gentherapist/requirements.txt
Django==4.2.7
textblob==0.17.1
python-decouple==3.8
